requirements.txt
Django==5.1
mysqlclient
django-ninja==1.3.0
django-ninja-extra==0.21.4
